requirements.txt
streamlit==1.32.2
torch==2.2.2
torchvision==0.17.2
timm==0.9.16
pillow==10.2.0
numpy==1.26.4
requests==2.31.0